# desktop_app/src/backend/models/chat/index.ts
import { asc, desc, eq } from 'drizzle-orm';
import { z } from 'zod';

import db from '@backend/database';
import { SelectChatSchema, chatsTable } from '@backend/database/schema/chat';
import {
  SelectMessagesSchema as DatabaseMessageRepresentationSchema,
  messagesTable,
} from '@backend/database/schema/messages';

const TransformedMessageSchema = DatabaseMessageRepresentationSchema.extend({
  /**
   * NOTE: id is stored in the database as a number, but we will convert it to a string as such:
   * `${chat.sessionId}-${message.id}`
   */
  id: z.string(),
});

export const ChatWithMessagesSchema = SelectChatSchema.extend({
  messages: z.array(TransformedMessageSchema),
});

type DatabaseMessage = z.infer<typeof DatabaseMessageRepresentationSchema>;
type TransformedMessage = z.infer<typeof TransformedMessageSchema>;
type Chat = z.infer<typeof SelectChatSchema>;
type ChatWithMessages = z.infer<typeof ChatWithMessagesSchema>;

export default class ChatModel {
  static generateCompositeMessageId = (chat: Chat, message: DatabaseMessage): string =>
    `${chat.sessionId}-${message.id}`;

  static async getAllChats(): Promise<ChatWithMessages[]> {
    const rows = await db
      .select()
      .from(chatsTable)
      .leftJoin(messagesTable, eq(chatsTable.id, messagesTable.chatId))
      .orderBy(desc(chatsTable.createdAt), asc(messagesTable.createdAt)); // Most recent chats first, then messages in order

    // Use a Map for better performance and type safety
    const chatMap = new Map<number, ChatWithMessages>();

    for (const row of rows) {
      const chat = row.chats;
      const message = row.messages;

      if (!chatMap.has(chat.id)) {
        chatMap.set(chat.id, {
          ...chat,
          messages: [],
        });
      }

      if (message) {
        // Parse the message content and add to the chat's messages array
        const parsedMessage = {
          ...message,
          id: `${chat.sessionId}-${message.id}`,
          content: JSON.parse(message.content as string),
        };
        chatMap.get(chat.id)!.messages.push(parsedMessage);
      }
    }

    // Convert Map to array and sort by createdAt to maintain order
    return Array.from(chatMap.values()).sort(
      (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
    );
  }

  static async getChatById(id: number): Promise<ChatWithMessages | null> {
    const rows = await db
      .select()
      .from(chatsTable)
      .leftJoin(messagesTable, eq(chatsTable.id, messagesTable.chatId))
      .where(eq(chatsTable.id, id))
      .orderBy(asc(messagesTable.createdAt)); // Order messages by creation time

    if (rows.length === 0 || !rows[0].chats) {
      return null;
    }

    const chat = rows[0].chats;
    const messages: TransformedMessage[] = [];

    for (const row of rows) {
      if (row.messages) {
        messages.push({
          ...row.messages,
          id: this.generateCompositeMessageId(chat, row.messages),
          content: JSON.parse(row.messages.content as string),
        });
      }
    }

    return {
      ...chat,
      messages,
    };
  }

  static async createChat(): Promise<ChatWithMessages> {
    // The sessionId is auto-generated by SQLite using the UUID expression
    // defined in the schema (see chat.ts schema file)
    const [chat] = await db
      .insert(chatsTable)
      .values({}) // No required fields, all handled by defaults
      .returning(); // SQLite returns the inserted row

    return {
      ...chat,
      messages: [],
    };
  }

  static async updateChat(id: number, data: { title?: string | null }): Promise<ChatWithMessages | null> {
    // First check if the chat exists
    const chat = await this.getChatById(id);
    if (!chat) {
      return null; // Will trigger 404 in the route handler
    }

    const [updatedChat] = await db
      .update(chatsTable)
      .set({
        title: data.title,
        // Manually update the timestamp since SQLite doesn't have ON UPDATE
        updatedAt: new Date().toISOString(),
      })
      .where(eq(chatsTable.id, id))
      .returning();

    return {
      ...updatedChat,
      messages: chat.messages, // Keep existing messages
    };
  }

  static async deleteChat(id: number): Promise<void> {
    // Note: Related chat_interactions will be cascade deleted
    // when that table is added (foreign key constraint)
    await db.delete(chatsTable).where(eq(chatsTable.id, id));
  }

  static async saveMessages(sessionId: string, messages: DatabaseMessage[]): Promise<void> {
    // First, find the chat by session ID
    const [chat] = await db.select().from(chatsTable).where(eq(chatsTable.sessionId, sessionId)).limit(1);

    if (!chat) {
      console.error(`Chat not found for session ID: ${sessionId}`);
      return;
    }

    // Clear existing messages for this chat to avoid duplicates
    await db.delete(messagesTable).where(eq(messagesTable.chatId, chat.id));

    // Save each message
    for (const message of messages) {
      await db.insert(messagesTable).values({
        chatId: chat.id,
        role: message.role,
        content: JSON.stringify(message),
      });
    }
  }
}